Criação do componente de modal
