+++ shared/src/components/domain/file/index.ts
@@ -1,0 +1,3 @@
+export { default as AttachmentDisplay } from './AttachmentDisplay';
+export { default as FilePreviewModal } from './FilePreviewModal';
+export { default as FileUpload } from './FileUpload';
