+++ shared/src/components/domain/index.ts
@@ -1,0 +1,3 @@
+export * from './file';
+export * from './transaction';
+export { default as BalanceCard } from './BalanceCard';
