+++ shared/src/components/domain/transaction/index.ts
@@ -1,0 +1,4 @@
+export { default as TransactionAdd } from './TransactionAdd';
+export { default as TransactionTypeBadge } from './TransactionTypeBadge';
+export { default as TransactionList } from './TransactionList';
+export { default as TransactionEdit } from './TransactionEdit';
