+++ shared/src/components/ui/index.ts
@@ -1,0 +1,4 @@
+export { default as Button } from './Button';
+export { default as Card } from './Card';
+export { default as Icon } from './Icon';
+export { default as ConfirmationModal } from './ConfirmationModal';
