+++ transactions-mfe/webpack.config.js
@@ -66,7 +66,9 @@ module.exports = {
         './TransactionsPage': './src/App.tsx'
       },
       remotes: {
-        shared: 'shared@http://localhost:3033/remoteEntry.js'
+        shared: 'shared@http://localhost:3033/remoteEntry.js',
+        dashboardMFE: 'dashboardMFE@http://localhost:3031/remoteEntry.js',
+        transactionsMFE: 'transactionsMFE@http://localhost:3032/remoteEntry.js',
       },
       shared: {
         react: {
